ExpressionViewer -> ValueViewer, factor out getBoxProps
